Revert "MULE-18063: Need a way to stop built-in schedule for CH apps (#8655)"
This reverts commit 569f5a44d133f1e92602e9e5aa9bf07d029dbf97.
