+++ core-tests/src/test/java/org/mule/runtime/core/internal/source/polling/DefaultSchedulerMessageSourceTestCase.java
@@ -4,40 +4,54 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.runtime.core.internal.source.scheduler;
+package org.mule.runtime.core.internal.source.polling;
 
 import static java.util.Collections.singletonMap;
 import static java.util.Optional.of;
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyLong;
+import static org.mockito.Mockito.doAnswer;
 import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.reset;
+import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
+import static org.mule.runtime.api.component.AbstractComponent.LOCATION_KEY;
 import static org.mule.runtime.api.component.location.ConfigurationComponentLocator.REGISTRY_KEY;
 import static org.mule.runtime.core.api.lifecycle.LifecycleUtils.disposeIfNeeded;
 import static org.mule.runtime.core.api.lifecycle.LifecycleUtils.stopIfNeeded;
 import static org.mule.tck.MuleTestUtils.APPLE_FLOW;
 import static org.mule.tck.MuleTestUtils.createFlowWithSource;
 import static org.slf4j.LoggerFactory.getLogger;
-
-import java.util.Map;
-
-import org.junit.After;
-import org.junit.Before;
 import org.mule.runtime.api.component.location.Location;
 import org.mule.runtime.api.exception.MuleException;
+import org.mule.runtime.api.scheduler.Scheduler;
+import org.mule.runtime.api.scheduler.SchedulerService;
 import org.mule.runtime.core.api.construct.Flow;
+import org.mule.runtime.core.api.event.CoreEvent;
 import org.mule.runtime.core.api.source.scheduler.FixedFrequencyScheduler;
 import org.mule.runtime.core.internal.context.DefaultMuleContext;
 import org.mule.runtime.core.internal.execution.MuleMessageProcessingManager;
 import org.mule.runtime.core.internal.policy.PolicyManager;
 import org.mule.runtime.core.internal.policy.SourcePolicy;
+import org.mule.runtime.core.internal.source.scheduler.DefaultSchedulerMessageSource;
+import org.mule.tck.SensingNullMessageProcessor;
 import org.mule.tck.junit4.AbstractMuleContextTestCase;
+import org.mule.tck.probe.PollingProber;
+import org.mule.tck.probe.Probe;
+
+import java.util.Map;
+import java.util.concurrent.atomic.AtomicReference;
+
+import org.junit.After;
+import org.junit.Before;
+import org.junit.Test;
 import org.slf4j.Logger;
 
 public class DefaultSchedulerMessageSourceTestCase extends AbstractMuleContextTestCase {
 
   private static final Logger LOGGER = getLogger(DefaultSchedulerMessageSourceTestCase.class);
   private PolicyManager policyManager;
-  protected SourcePolicy sourcePolicy;
+  private SourcePolicy sourcePolicy;
   private final String MESSAGE_PROCESSING_MANAGER_KEY = "_muleMessageProcessingManager";
 
   @Before
@@ -59,6 +73,62 @@ public class DefaultSchedulerMessageSourceTestCase extends AbstractMuleContextTe
     return singletonMap(REGISTRY_KEY, componentLocator);
   }
 
+  @Test
+  public void simplePoll() throws Exception {
+
+    DefaultSchedulerMessageSource schedulerMessageSource = createMessageSource();
+
+    SensingNullMessageProcessor flow = getSensingNullMessageProcessor();
+    schedulerMessageSource.setListener(flow);
+    schedulerMessageSource.setAnnotations(singletonMap(LOCATION_KEY, TEST_CONNECTOR_LOCATION));
+
+    doAnswer(invocationOnMock -> {
+      CoreEvent inputEvent = invocationOnMock.getArgument(0);
+      flow.process(inputEvent);
+      return null;
+    }).when(sourcePolicy).process(any(CoreEvent.class), any(), any());
+
+    schedulerMessageSource.trigger();
+    new PollingProber(RECEIVE_TIMEOUT, 100).check(new Probe() {
+
+      @Override
+      public boolean isSatisfied() {
+        return flow.event != null;
+      }
+
+      @Override
+      public String describeFailure() {
+        return "flow event never set by the source flow";
+      }
+    });
+  }
+
+  @Test
+  public void disposeScheduler() throws Exception {
+    SchedulerService schedulerService = muleContext.getSchedulerService();
+    reset(schedulerService);
+
+    AtomicReference<Scheduler> pollScheduler = new AtomicReference<>();
+
+    doAnswer(invocation -> {
+      Scheduler scheduler = (Scheduler) invocation.callRealMethod();
+      pollScheduler.set(scheduler);
+      return scheduler;
+    }).when(schedulerService).cpuLightScheduler();
+
+    DefaultSchedulerMessageSource schedulerMessageSource = createMessageSource();
+    verify(schedulerService).cpuLightScheduler();
+
+    schedulerMessageSource.start();
+
+    verify(pollScheduler.get()).scheduleAtFixedRate(any(), anyLong(), anyLong(), any());
+
+    schedulerMessageSource.stop();
+    schedulerMessageSource.dispose();
+
+    verify(pollScheduler.get()).stop();
+  }
+
   private DefaultSchedulerMessageSource schedulerMessageSource;
 
   @After
@@ -67,7 +137,7 @@ public class DefaultSchedulerMessageSourceTestCase extends AbstractMuleContextTe
     disposeIfNeeded(schedulerMessageSource, LOGGER);
   }
 
-  protected DefaultSchedulerMessageSource createMessageSource() throws Exception {
+  private DefaultSchedulerMessageSource createMessageSource() throws Exception {
     schedulerMessageSource =
         new DefaultSchedulerMessageSource(muleContext, scheduler(), false);
     schedulerMessageSource.setAnnotations(getAppleFlowComponentLocationAnnotations());
@@ -88,4 +158,5 @@ public class DefaultSchedulerMessageSourceTestCase extends AbstractMuleContextTe
     factory.setFrequency(1000);
     return factory;
   }
+
 }
