+++ core/src/main/java/org/mule/runtime/core/internal/source/scheduler/DefaultSchedulerMessageSource.java
@@ -6,11 +6,8 @@
  */
 package org.mule.runtime.core.internal.source.scheduler;
 
-import static java.lang.Boolean.parseBoolean;
-import static java.lang.System.getProperty;
 import static java.util.Collections.emptyList;
 import static java.util.Optional.empty;
-import static org.mule.runtime.api.util.MuleSystemProperties.MULE_DISABLE_SCHEDULERS;
 import static org.mule.runtime.core.api.config.i18n.CoreMessages.failedToScheduleWork;
 import static org.mule.runtime.core.api.source.MessageSource.BackPressureStrategy.FAIL;
 import static org.mule.runtime.core.api.util.ClassUtils.withContextClassLoader;
@@ -66,7 +63,6 @@ public class DefaultSchedulerMessageSource extends AbstractComponent
 
   private final PeriodicScheduler scheduler;
   private final boolean disallowConcurrentExecution;
-  private static boolean disabled = parseBoolean(getProperty(MULE_DISABLE_SCHEDULERS));
 
   private Scheduler pollingExecutor;
   private ScheduledFuture<?> schedulingJob;
@@ -101,7 +97,7 @@ public class DefaultSchedulerMessageSource extends AbstractComponent
 
   @Override
   public synchronized void start() throws MuleException {
-    if (started || disabled) {
+    if (started) {
       return;
     }
     try {
@@ -130,9 +126,7 @@ public class DefaultSchedulerMessageSource extends AbstractComponent
 
   @Override
   public void trigger() {
-    if (!disabled) {
-      pollingExecutor.execute(() -> withContextClassLoader(muleContext.getExecutionClassLoader(), () -> poll()));
-    }
+    pollingExecutor.execute(() -> withContextClassLoader(muleContext.getExecutionClassLoader(), () -> poll()));
   }
 
   @Override
@@ -206,9 +200,6 @@ public class DefaultSchedulerMessageSource extends AbstractComponent
    */
   @Override
   public void initialise() throws InitialisationException {
-    if (disabled) {
-      return;
-    }
     getFromAnnotatedObject(componentLocator, this)
         .ifPresent(flow -> this.flowConstruct = flow);
 
@@ -284,15 +275,5 @@ public class DefaultSchedulerMessageSource extends AbstractComponent
     public FlowConstruct getFlowConstruct() {
       return flowConstruct;
     }
-  }
-
-  /**
-   * This method is only for testing proposes.  Don't use.
-   *
-   * @param disabled
-   * @deprecated
-   */
-  static void setDisabled(boolean disabled) {
-    DefaultSchedulerMessageSource.disabled = disabled;
   }
 }
